my_bot/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
certifi==2025.1.31
click==8.1.8
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==21.11.1
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
Telethon==1.39.0
typing_extensions==4.12.2
tzlocal==5.3
uvicorn==0.34.0
